requirements.txt
# Data Collection
python-binance==1.0.19
praw==7.7.1
pycoingecko==3.1.0

# Message Queue & Stream Processing
kafka-python==2.0.2
apache-flink==1.17.0

# Storage
minio==7.2.0

# ML & Analytics
scikit-learn==1.3.2
tensorflow==2.11.0
torch==2.1.1
nltk==3.8.1
spacy==3.7.2
pandas==1.3.0
numpy==1.21.4
pandas-ta==0.3.14b

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Monitoring
prometheus-client==0.19.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.2 